Move bcrypt hash to threadpool outside of async
Before this commit, bcrypt hashes were being calculated in an
async context, blocking the async state machine.
Moving computation into `spawn_blocking` is not enough, because if
too many requests arrive at the same time, we are at the mercy of
our operating system's scheduler. If we get 100 requests at the
same time, they might all end roughly at the same time instead of
some of them ending as soon as possible. To remedy this, a semaphore
is required, allowing only some of the requests to proceed. This
should lower total request latency time.